Обновил сервисы логики по работе с сущностями из-за нового типа

+++ src/main/java/ru/severstal/notes/repository/NoteRepository.java
@@ -1,7 +1,7 @@
 package ru.severstal.notes.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import ru.severstal.notes.model.Note;
+import ru.severstal.notes.entity.Note;
 
-public interface NoteRepository extends CrudRepository<Note, Integer> {
+public interface NoteRepository extends CrudRepository<Note, Long> {
 }
